Create `ReviewHeader` component
